Pass down API version in SimpleJSONRPCServer.
JSON-RPC version 1.0 requires the 'error' key to be set in the response
even if there is no error.  This is correctly implemented
Payload.response -- but for that to work, the API version has to be
passed down in SimpleJSONRPCServer when it calls dumps to construct the
response.  This was missing, so that JSON-RPC 1.0 requests were answered
with a 2.0 response.